Crate burrito

Burrito is a monadic IO interface, similar to Haskell's IO monad.

A single Burrito type wraps a single IO handle, whether its the stdio, a file on the system, or a socket address. There are several ways to create a Burrito.

burrito() creates an stdio Burrito.

Burrito::wrap() wraps a Result, where T is an IO handle.

Burrito::wrap_func() wraps a function which returns a Result, where T is an IO handle.

Types which implement FromPath can be wrapped using Burrito::from_path()

Types which implement FromAddr can be wrapped using Burrito::from_addr()

IO actions can be performed directly on the Burrito; the result of these actions can be accessed using the and_then method. IO failure can be handled using the or_else method.

As a simple example of a Burrito in action, this code will echo stdin on stdout.

use burrito::burrito;

burrito().read_line().and_then(|echo, burrito| burrito.print_line(&echo));

The RealWorld type

The default Burrito, which is returned by burrito(), wraps a type called RealWorld. This encapsulates the handles to stdin, stdout, stderr. It does not lock these handles; it will attempt to lock them each time they are written to / read from.

The RealWorld type is not directly accessible except wrapped inside of the Burrito returned by burrito().

Using the monadic interface

Every Burrito has two type parameters. The T type is the type of the IO handle it wraps. The A type, however, is the type of the data returned by the most recent IO call.

All of the IO calls implemented on Burrito consume the object and return a new Burrito; if that call returned data of some kind, that data is stored inside that new Burrito.

and_then()

The prefered way to access the data inside a burrito is the and_then method. and_thenn takes a function of A and Burrito<(), T> to any Burrito, and applies that function to the A data inside the Burrito and a new Burrito wrapped around the IO handle.

A common pattern when using Burritos is a back and forth between IO methods and and_then calls, possibly including IO methods on the burrito and the end of the and_then.

use burrito::burrito;

burrito().read_line().and_then(|input, burrito| {
    let output = // Generate output based on the input in some way.
    burrito.print_line(output)
}).read_line().and_then(|input, burrito| {
    // Do more stuff with the next input.
    burrito
});

Note that the function passed to and_then is only called in the event that the Burrito is not in a state of failure.

and()

The and() method is less powerful than the and_then() method because it does not provide access to the data inside the Burrito. Its main use is to close this IO handle when its done being used and to open another one, which could possibly be of a different type. It is not lazy, and the IO handle will actually be opened even if the Burrito is in a state of failure.

or_else()

The or_else() method provides access to the inner error in the event that the Burrito represents an IO handle in a state of failure. As soon as any IO method on that handle returns a failure, the Burrito remains in a state of failure until re-instantiated. This method must either create a Burrito of the same type (presumably replacing the original Burrito) or else diverge, e.g. by exiting or panicking.

or()

The or() method enables replacing the Burrito with another of the same types, but does not provide access to the inner error. It is not lazy, and will actually open the handle even if the Burrito is not in a state of failure.

Structs

Burrito
The fundamental monadic type of the burrito library.

Traits

FromAddr

FromPath

Functions

burrito
Create a default burrito (wrapping the stdio handles).