Enum burrito::Burrito

pub enum Burrito<T> {
    Good(T),
    Bad(Error),
}

The Burrito enum provides a monadic interface to an IO handle. They encapsulate all the error and state in the messy real world so that your program can exist in a state of cheesey, carne asada goodness. Or something. I don't know; it made sense for a second. Let me find a blog post.

Variants

Good

Bad

Methods

impl<T> Burrito<T>

fn new(t: T) -> Burrito<T>

Wrap a tortilla around some ucky stateful thing.

Burrito::new() can take any type at all, but probably its better for you to use a type that implements Read or Write.

fn and_then<F, U>(self, f: F) -> Burrito<U> where F: FnOnce(T) -> Burrito<U>

If the burrito is tasty, apply f() to its yummy innards. If its gone bad, the same burrito will be returned.

fn or_else<F>(self, f: F) -> Burrito<T> where F: FnOnce(Error) -> Burrito<T>

If the burrito is bad, create a new burrito by apply f() to the bad stuff inside. If its still good, the same burrito will be returned.

impl Burrito<File>

fn new_from_path<P: AsRef<Path>>(path: P) -> Burrito<File>

Wrapping an existing IO handle can be good, but you still have to deal with the error-prone nature of instantiating an IO handle. This will open a file handle (or not) inside of the burrito.

impl<T> Burrito<T> where T: Read

fn read(self, buf: &mut Vec<u8>) -> Burrito<T>

Read from inside the burrito. This will call read_to_end() on the inner readable, not read().

fn read_to_string(self, buf: &mut String) -> Burrito<T>

Read from inside the burrito to a string.

impl<T> Burrito<T> where T: Write

fn write(self, buf: &[u8]) -> Burrito<T>

Write from inside the burrito. This will call write_all() on the inner writeable, not write().

fn write_fmt(self, fmt: Arguments) -> Burrito<T>

Write a formatted String from inside the burrito.

impl Burrito<RealWorld>

fn write_to_err(self, buf: &[u8]) -> Burrito<RealWorld>

Write to stderr rather than stdout.

fn write_fmt_to_err(self, fmt: Arguments) -> Burrito<RealWorld>

Write a formatted String to stderr.