Struct burrito::Burrito

pub struct Burrito<A, T>(_);

The fundamental monadic type of the burrito library.

Burrito implements different IO methods depending on the traits implemented by the IO handle it wraps. These methods have the same name as the methods associated with that trait, though their signature differs somewhat.

Methods

impl<T> Burrito<(), T>

These methods construct Burritos of arbitrary types.

fn wrap(inner: Result<T>) -> Burrito<(), T>

The basic constructor for Burrito. This takes an io::Result<T>, where T is the type being wrapped by the Burrito, io::Result<T> is the return type of the constructor for most IO handle types.

use std::fs::File;
use burrito::Burrito;

let burrito = Burrito::wrap(File::create("/foo/bar/baz"));

fn wrap_func<F: FnOnce() -> Result<T>>(f: F) -> Burrito<(), T>

A constructor for Burrito which takes a function and wraps the result of that function.

use std::fs::{self, File};
use burrito::Burrito;

let burrito = Burrito::wrap_func(|| {
    try!(fs::metadata("/foo/bar/baz"));
    File::open("/foo/bar/baz")
});

impl Burrito<(), ()>

These two functions are constructors for types which can be constructed from paths and socket addresses.

fn from_path<P: AsRef<Path>, T: FromPath>(path: P) -> Burrito<(), T>

Constructs an IO handle using the path argument, according to that IO handle's implementation of FromPath, then wraps that handle in a Burrito. It is a good idea to type annotate this call to ensure the correct kind of handle is constructed.

use std::fs::File;
use burrito::{Burrito, FromPath};

let burrito = Burrito::from_path::<_, File>("/foo/bar/baz");

fn from_addr<A: ToSocketAddrs, T: FromAddr>(addr: A) -> Burrito<(), T>

Constructs an IO handle using the addr argument, according to that IO handle's implementation of FromAddr, then wraps that handle in a Burrito. It is a good idea to type annotate this call to ensure the correct kind of handle is constructed.

use std::net::TcpStream;
use burrito::{Burrito, FromAddr};

let burrito = Burrito::from_addr::<_, TcpStream>("localhost:12345");

impl<A, T> Burrito<A, T>

These methods are defined for all Burritos.

fn and<B, U>(self, alternative: Burrito<B, U>) -> Burrito<B, U>

Allows you to 'pivot' to a new Burrito if this one is good, or to remain in a state of failure if this Burrito has failed. See the module level documentation for more info.

fn and_then<B, U, F>(self, f: F) -> Burrito<B, U> where F: FnOnce(A, Burrito<(), T>) -> Burrito<B, U>

Allows access to data returned by the most recent IO call on this Burrito; this function must return another Burrito of some kind or else diverge. See the module level documentation for more info.

fn or(self, alternative: Burrito<A, T>) -> Burrito<A, T>

Allows you to substitute this Burrito for another of the same type if it has gone bad.

fn or_else<F>(self, f: F) -> Burrito<A, T> where F: FnOnce(Error) -> Burrito<A, T>

Allows access to the error thrown if this Burrito has gone bad. This function must return another Burrito of the same type or else diverge. See the module level documentation for more info.

fn ignore(self) -> Burrito<(), T>

Drops any data returned by the most recent IO call.

fn is_good(&self) -> bool

Returns true if the Burrito has not failed.

fn is_bad(&self) -> bool

Returns true if the Burrito has failed.

fn ok(self) -> Result<(A, T)>

Converts the Burrito to a Result of both the handle and the most recently returned data.

fn to_data(self) -> Result<A>

Converts the Burrito to a Result of the most recently returned data.

fn to_handle(self) -> Result<T>

Converts the Burrito to a Result of the IO handle wrapped within.

impl<A, T> Burrito<A, T> where T: Read

fn read(self, n: usize) -> Burrito<Vec<u8>, T>

Performs a read on the IO handle inside the burrito. Will read into a buffer of n bytes.

Though the buffer passed to Read::read() can be stack allocated, this function allocates the buffer on the heap, so that its length can be determined by the function call. The Vec<u8> returned by this type will contain all of the bytes read from the call; if that is less than n, it will not include any null bytes.

fn read_to_end(self) -> Burrito<Vec<u8>, T>

Reads to the end of the handle inside the burrito, returning a Vec<u8> of bytes.

fn read_to_string(self) -> Burrito<String, T>

Reads everything from the handle into a String.

impl<A, T> Burrito<A, T> where T: Write

fn write(self, buf: &[u8]) -> Burrito<usize, T>

Writes the content of the buf to the IO handle; returns the number of bytes written.

fn write_all(self, buf: &[u8]) -> Burrito<(), T>

Writes the content of the buf to the IO handle; will write all of the bytes unless it fails.

fn write_fmt(self, buf: Arguments) -> Burrito<(), T>

Writes formatted text to the IO handle.

impl<A, T> Burrito<A, T> where T: Seek

fn seek(self, pos: SeekFrom) -> Burrito<u64, T>

Seeks to a position in the IO handle; returns the actual position that has been seeked to.

impl<A, T> Burrito<A, T> where T: BufRead

fn fill_buf(self) -> Burrito<(), T>

Fills the buffer on the buffered reader. Unlike the underlying fill_buf macro, this does not return a reference to the bytes in the buffer.

fn consume(self, amt: usize) -> Burrito<(), T>

Marks amt bytes in the buffer as consumed.

fn read_until(self, byte: u8) -> Burrito<Vec<u8>, T>

Reads from the buffered reader until the byte is reached.

fn read_line(self) -> Burrito<String, T>

Reads a line from the buffered reader.

fn split(self, byte: u8) -> Result<Split<T>>

Generates a Split Iterator of the underlying buffered reader. This will be wrapped in a result because the IO handle may have failed at some point in the past.

fn lines(self) -> Result<Lines<T>>

Generates a Lines Iterator of the underlying buffered reader. This will be wrapped in a result because the IO handle may have failed at some point in the past.

impl<A> Burrito<A, RealWorld>

These methods are implemented only for the stdio Burrito. Note that RealWorld implements both Read and Write, and so the stdio Burrito also has all methods for Burritos wrapping handles which implement those traits; the methods associated with the Write trait write to stdout, whereas a set of special to_err() methods write to stderr.

fn print_line(self, buf: &str) -> Burrito<(), RealWorld>

Prints a string to stdout, with a newline affixed to the end. Internally, it calls write_all; to use it like the println!() macro, you can use a reference to a format macro - that is &format!().

use burrito::burrito;

burrito().print_line(&format!("2 + 2 = {}", 4));

fn read_line(self) -> Burrito<String, RealWorld>

Reads a line from stdin. This has the same behavior as the read_line() method on io::Stdin.

use burrito::burrito;
 
let input = burrito().read_line().to_data();

fn write_to_err(self, buf: &[u8]) -> Burrito<usize, RealWorld>

Performs a write to stderr instead of stdout.

fn write_all_to_err(self, buf: &[u8]) -> Burrito<(), RealWorld>

Performs a write_all to stderr instead of stdout.

fn write_fmt_to_err(self, fmt: Arguments) -> Burrito<(), RealWorld>

Performs a write_fmt to stderr instead of stdout.